+++ src/services/color/eui_palettes.js
@@ -1,0 +1,77 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export const palettes = {
+  color_blind: {
+    colors: [
+      '#1ea593',
+      '#2b70f7',
+      '#ce0060',
+      '#38007e',
+      '#fca5d3',
+      '#f37020',
+      '#e49e29',
+      '#b0916f',
+      '#7b000b',
+      '#34130c',
+    ],
+  },
+  color_blind_15: {
+    colors: [
+      '#000000',
+      '#004949',
+      '#009292',
+      '#ff6db6',
+      '#ffb6db',
+      '#490092',
+      '#006ddb',
+      '#b66dff',
+      '#6db6ff',
+      '#b6dbff',
+      '#920000',
+      '#924900',
+      '#db6d00',
+      '#24ff24',
+      '#ffff6d',
+    ]
+  },
+  eui_colors: {
+    colors: [
+      '#0079A5',
+      '#017F75',
+      '#E5830E',
+      '#A30000',
+      '#DD0A73',
+    ]
+  },
+  eui_colors_light: {
+    colors: [
+      '#009ED8',
+      '#01B2A4',
+      '#F39B33',
+      '#D60000',
+      '#F5258C',
+    ]
+  },
+  spectrum: {
+    colors: [
+      '#882E72',
+      '#B178A6',
+      '#D6C1DE',
+      '#1965B0',
+      '#5289C7',
+      '#7BAFDE',
+      '#4EB265',
+      '#90C987',
+      '#CAE0AB',
+      '#F7EE55',
+      '#F6C141',
+      '#F1932D',
+      '#E8601C',
+      '#DC050C',
+    ]
+  }
+};